requirements.txt
dnspython==2.4.2
Pillow==10.1.0
pymongo==4.6.0
python-dateutil==2.8.2
six==1.16.0
